requirements.txt
fastapi==0.65.1
FastAPI-SQLAlchemy==0.2.1
pydantic==1.8.2
alembic==1.6.3
psycopg2==2.8.6
uvicorn[standard]==0.13.4
pytest==6.2.4
requests==2.25.1
typer==0.3.2
